refactor(blind_spot): move util functions outside of class
