+++ planning/behavior_velocity_planner/autoware_behavior_velocity_blind_spot_module/include/autoware/behavior_velocity_blind_spot_module/scene.hpp
@@ -15,17 +15,13 @@
 #ifndef AUTOWARE__BEHAVIOR_VELOCITY_BLIND_SPOT_MODULE__SCENE_HPP_
 #define AUTOWARE__BEHAVIOR_VELOCITY_BLIND_SPOT_MODULE__SCENE_HPP_
 
+#include <autoware/behavior_velocity_blind_spot_module/util.hpp>
 #include <autoware/behavior_velocity_planner_common/scene_module_interface.hpp>
-#include <autoware/behavior_velocity_planner_common/utilization/boost_geometry_helper.hpp>
 #include <autoware/behavior_velocity_planner_common/utilization/state_machine.hpp>
 #include <rclcpp/rclcpp.hpp>
 
-#include <autoware_perception_msgs/msg/predicted_object.hpp>
 #include <autoware_perception_msgs/msg/predicted_objects.hpp>
-#include <geometry_msgs/msg/point.hpp>
-#include <tier4_planning_msgs/msg/path_with_lane_id.hpp>
 
-#include <lanelet2_core/LaneletMap.h>
 #include <lanelet2_routing/RoutingGraph.h>
 
 #include <memory>
@@ -36,21 +32,6 @@
 
 namespace autoware::behavior_velocity_planner
 {
-
-/**
- * @brief  wrapper class of interpolated path with lane id
- */
-struct InterpolatedPathInfo
-{
-  /** the interpolated path */
-  tier4_planning_msgs::msg::PathWithLaneId path;
-  /** discretization interval of interpolation */
-  double ds{0.0};
-  /** the intersection lanelet id */
-  lanelet::Id lane_id{0};
-  /** the range of indices for the path points with associative lane id */
-  std::optional<std::pair<size_t, size_t>> lane_id_interval{std::nullopt};
-};
 
 /**
  * @brief represent action
@@ -81,8 +62,6 @@ using BlindSpotDecision = std::variant<InternalError, OverPassJudge, Unsafe, Saf
 class BlindSpotModule : public SceneModuleInterface
 {
 public:
-  enum class TurnDirection { LEFT, RIGHT };
-
   struct DebugData
   {
     std::optional<geometry_msgs::msg::Pose> virtual_wall_pose{std::nullopt};
@@ -146,12 +125,6 @@ private:
   void reactRTCApprovalByDecision(
     const Decision & decision, tier4_planning_msgs::msg::PathWithLaneId * path);
 
-  std::optional<InterpolatedPathInfo> generateInterpolatedPathInfo(
-    const tier4_planning_msgs::msg::PathWithLaneId & input_path) const;
-
-  std::optional<lanelet::ConstLanelet> getSiblingStraightLanelet(
-    const std::shared_ptr<const PlannerData> planner_data) const;
-
   /**
    * @brief Generate a stop line and insert it into the path.
    * A stop line is at an intersection point of straight path with vehicle path
@@ -187,33 +160,6 @@ private:
     const lanelet::ConstLanelets & blind_spot_lanelets,
     const geometry_msgs::msg::Pose & stop_line_pose, const lanelet::CompoundPolygon3d & area,
     const double ego_time_to_reach_stop_line);
-
-  /**
-   * @brief Create half lanelet
-   * @param lanelet input lanelet
-   * @return Half lanelet
-   */
-  lanelet::ConstLanelet generateHalfLanelet(const lanelet::ConstLanelet lanelet) const;
-
-  lanelet::ConstLanelet generateExtendedAdjacentLanelet(
-    const lanelet::ConstLanelet lanelet, const TurnDirection direction) const;
-  lanelet::ConstLanelet generateExtendedOppositeAdjacentLanelet(
-    const lanelet::ConstLanelet lanelet, const TurnDirection direction) const;
-
-  lanelet::ConstLanelets generateBlindSpotLanelets(
-    const tier4_planning_msgs::msg::PathWithLaneId & path) const;
-
-  /**
-   * @brief Make blind spot areas. Narrow area is made from closest path point to stop line index.
-   * Broad area is made from backward expanded point to stop line point
-   * @param path path information associated with lane id
-   * @param closest_idx closest path point index from ego car in path points
-   * @return Blind spot polygons
-   */
-  std::optional<lanelet::CompoundPolygon3d> generateBlindSpotPolygons(
-    const tier4_planning_msgs::msg::PathWithLaneId & path, const size_t closest_idx,
-    const lanelet::ConstLanelets & blind_spot_lanelets,
-    const geometry_msgs::msg::Pose & stop_line_pose) const;
 
   /**
    * @brief Check if object is belong to targeted classes
